+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add breakpoint `condition` parameter by [@alexruperez](https://github.com/alexruperez).
 - Support Xcode Extension product type https://github.com/xcodeswift/xcproj/pull/190 by @briantkelley
 - Support for the legacy Build Carbon Resources build phase https://github.com/xcodeswift/xcproj/pull/196 by @briantkelley
+- Support for custom build rules by https://github.com/xcodeswift/xcproj/pull/197 @briantkelley
 
 ## 1.7.0